Polishing Reporter
Reporter should now be fully able to generate a report on a student. Still not able to get tutor session data, but if a fully realized student object is passed, most of its data will be formatted and reported. Still have to make the log class/file and do appropriate error handling.

I also added some new private variables to student to hold the data Reporter reports on. May want to reorganize this data later.